+++ LAB/Models/Raw.cs
@@ -22,9 +22,9 @@ namespace LAB.Models
         [Required]
         public int MeasurementId { get; set; }
 
-        public int Sum { get; set; }
+        public double Sum { get; set; }
         
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
 
        public IEnumerable<Ingredients> Ingredients { get; set; }
