+++ LAB/Models/Sell.cs
@@ -13,8 +13,8 @@ namespace LAB.Models
         public int FinishedProductsId { get; set; }
         public FinishedProducts FinishedProducts { get; set; }
         public DateTime Date { get; set; } = DateTime.Now;
-        public int Quantity { get; set; }
-        public int Sum { get; set; }
+        public double Quantity { get; set; }
+        public double Sum { get; set; }
         public int EmployeeId { get; set; }
         public Employee Employee { get; set; }
     }
